almost fairly accurate rendering.
If anyone has experience with framebuffer <-> texture formats that produce weird:  almost-accurate-but-not-quite-componding-feedback-texture errors, please leave me a comment; and see the flame menu program. I'm pretty sure it's a data-format issue where the sampler is off by like... a fraction of a texel. And I don't know how to fix it.