FINALLY!!!! Fixed the font glyphs- using SDL.
Something prevented the glyph texture from completely loading; but it was partially loading. So, there's more to the backend that's not obvious..... AT ALL!  Some useful tools:  NVIDIA Texture Tools, SDL (which got it first try), and wikipedia.

I implemented part of the BMP specification and loaded the data that way - which was thoroughly tried with:  glTexImage2D, glTextureImage2D, glTextureImage2DEXT, etc... SDL seems to load the data into the GL backend, finally, but there are some inconsistencies with format; and that is the real issue. Documentation on OpenGL is VERY SPARSE. Please leave comments if you're more familiar with the framework(s). Also, some more thorough information about loading BMP data would be appreciated. I found that some programs don't save the header data properly; and there's no way to recover the data for the image width / height. So, there must be some assumptions being made that are platform specific; but I don't see how you can reconstruct an image if you don't have at least those quantities. (I've tried BITMAPCOREHEADER, BITMAPV5HEADER, BITMAPINFOHEADER, and played with the results looking for the lost width / height of the color image data; but even paint.net wasn't saving it properly!) (or something is off; but I ran out of options... GEEZ!)  The NVIDIA tool details the BMP internal format that IT wants; but it doesn't fill out the header data - except for graphics related specific headers (which I stayed away from... this is a civilian application)