couple more issues surrounding:  std::filesystem, sorting, heap allocation, and const &-ing
Added binary sorting algorithm to simpleList. For the most part we're using copying heavily. So, there's issues when treating data on the stack - and especially when we're not sure about our simpleString copy constructors - because of other un-forseen issues with taking in std::string const char*'s. I'm pretty sure we're causing most of them; and have gone through lengths to fix them. So, the array, empty string issue could be treated much more easily if we knew that malloc was returning the correct number of chars (!!!) for one... The rest I'm sure will work. But, please remember to create copy constructors and override operator= because the container classes will lose your memory and you'll hit the wall >_<