Cleaned up a nasty compiler error:  brogueView didn't like the brogueUIBuilder.
The static functions weren't compiling no matter where I tried putting the getPaddedBoundary() function. So, to avoid further headaches, I just rewrote it in the brogueUIData class until all that code gets situated. There's probably a header "inferred" by the compiler which makes it near impossible to fix the 1000 errors that pop up out of nowhere. So, it's better to take some time to prune the headers, once in a while.