Think we finally have a render / control loop that will work.
Simple Goals:  1) Create a UI framework for the rest of brogueHd. 2) Handle generic UI controls, 3) Handle the main control loop to transition them in real time. This one was fun to set up! Give it a try if you want to play with a very simple opengl control library for a rouguelike. Current controls:  Button, Text, and ListView (with scrolling)