Created Event Controller:  using std::bind
UI Events now have a path to not-so-irritating-UI-tree-issues. It's apparently much cleaner to forward events straight from the event site. This was apparent after playing with the render loop and the UI tree. So, there's a clever way to build an "event aggregator" but you have to pre-define the event class types. See eventController.h. This implementation is much cleaner; and should be the way to finish this "UI Framework" up. Maybe even a pretty good substitute for libtcod - for more experienced UI developers.