Refactoring:  brogueProgramContainer, brogueMainProgram
Found a problem with switching programs on the render thread. The create program was returning NULL, so the code is now reorganized to put the entire program on the GPU at startup. There won't be any context switching for GLFW, or GL, or any other backend component. Just some refactoring of the rendering components to keep things in order.