Finally, figured out our buffer feedback issue. PART 1. (of Dante's Inferno >_<)
OpenGL Issue:  Framebuffer feedback. Or, texture feedback. Several problems occurred with various solutions.. just not the simple and obvious one.

Error:  Pixel precision SEEMED to be the issue. The "green tearing" of the texture when sampled from another spot in the texture space was a nasty issue - especially when we weren't sure about the data format (!). I'm not convinced it's correct, yet, because there's a sampling error that seems to compoud up-the-screen while we're feeding back the buffer.

Solution:  (Part 1) was to call glFlush() after the previous render calls. I'm not sure why this wasn't more of an issue - it should've been completely nasty on the screen - not just at the edges; and not just for SOME of the coordinates!

Next part, figure out what the exact image format it (if it's off at all!), also, find out why NVIDIA isn't supporting glNamedBufferSubCopyData (propery!). This feedback would've been settled by a simple VBO copy (!!!!!) It was returning GL_INVALID_VALUE.