Trying to get rid of frame buffer "chatter":  made progress on separating out by texture
The problem is that the sampling of the blue flames in the title screen is noisy. Texel fetch cleared up a big error; but there's still a sampling problem overlaying the blue and red view streams. So, going to check this in; and try drawing the textures one at a time.