Finally finished fixing the mask texture (see description)
The problem was a combination of texel precision error + framebuffer "thrashing" (or something like that). The data for the flame background was supposed to be used in a feedback loop. This is ideal for rendering on the GPU; and leaving the texture on the frame buffer. This will work until you mask off the data you just drew. So, applying a title mask over the flame background was problematic. Things worked much better once they were all separated into component textures - one per UI element. So, the whole "UI library" operates on an array of textures; and one "mixing" frame program. The other issue was getting the glFlush sequence correct. Generally, flushing was a good idea - especially to fix the "texel precision error" - which produced green pixels where they weren't supposed to be. So, the GL rendering sequence has trouble texture sampling (for this feedback loop) when there's other things going on - using the frame buffer. (Any experience here is appreciated. Please leave comments)