UI State Changes:  Framework for sequencing added, also added the first animator for the UI - the game log
State changes for the UI elements were needed to handle animation of the game log. So, there's a simple framework to handle them during the render loop. Three new functions were added to the brogueViewBase tree; and consumed to have controller methods that can monitor them over the course of state changes. So, the game log has it's animation handled while the render loop is aware of everything else happening on the UI tree.