Trying new tiling method:  the layout will first be tiled before worrying about planting gridRegion rooms
The accretion "bolt-on" idea is still what will be used; but the tiling will be easier working with gridRect instances; and avoiding iterating the grid. So, the center of the next tile is going to be moved as close to the entrance room as possible - using a combination of methods - including random placement to shorten the length of iteration.