Some success on the rectangle packing front; and some issues.
Rectangle packing: Not yet ideal just using a simple heuristic check. But, there might be better uses of this once connection points for the room graph are in place. Probably "pulling" a room closer to the door; but might not be needed.

Largest sub-rectangle routine:  This starts a better path towards success. "get the largest area sub-rect" from a space. Now, tile it with some random rect. Repeat until it's all done; and try utilizing our packing routine (optional). The rest of the space can be found using flood fill.