Refactored:  simpleHash -> simpleHash + simpleHashCore
The core now has the "Basic" and "Std" versions. There was hashing issue - which prompted me to try the unordered map instead of our "simple" one. So, the facade simpleHash is still in place. The "Basic" core is still there for fun; and can be worked out later. There were a few bugs in the hasher that were fixed. Hopefully, this is the final version; but more learning on SNIAF is used frequently to help aid polymorphic calls and variadic methods of solving these typical design issues.