Went through old BST code, cleaned it up for writing unit tests. Also, copied in boost's hash_combine - which is now operating
There were hash collisions for char arrays. I'm guessing my puny x = x ^ (hash << 1) didn't do enough to avoid the 255 values it was receiving from the char array. boost's is likely a better choice. It was a simple copy-paste function; but there could be a boost library in this code's future...