transferring layout code from another project...
see layoutGenerator.h for "the plan" ... This design was for another rogue project. The basic steps are described in the comments. The problem was to create a very robust layout cycle for:  room layout, triangulation, terrain, layout dijkstra's map issues, repeatable results, declarative syntax, interfacing, etc... There is heavy usage of Dijkstra's algorithm for the rest of the brogue game. So, this is essentially carving out the (reusable) robust code segments for utilizing these algorithms on the fly.